dbs/qdrant/requirements.txt
qdrant-client>=1.1.4
optimum[onnxruntime]==1.8.2
transformers==4.28.1
sentence-transformers==2.2.2
pydantic>=1.10.7, <2.0.0
fastapi>=0.95.0, <1.0.0
httpx>=0.24.0
aiohttp>=3.8.4
uvicorn>=0.21.0, <1.0.0
python-dotenv>=1.0.0
srsly>=2.4.6